google-adk/requirements.txt
google-adk>=1.25.0
notion-client>=2.2.0
gspread>=6.0.0
google-auth>=2.25.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
